# backend/src/swagger/userRoutes.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Backend API Documentation",
    "version": "1.0.0",
    "description": "API documentation for the Resume Builder Application",
    "contact": {
      "email": "guptaabhay.75way@gmail.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development server"
    }
  ],
  "paths": {
    "/users/register": {
      "post": {
        "summary": "Register a new user",
        "operationId": "createUser",
        "tags": ["User"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "User registration data",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string", "example": "abhay" },
                  "email": { "type": "string", "example": "xyz@gmail.com" },
                  "password": { "type": "string", "example": "abhay" },
                  "confirmPassword": { "type": "string", "example": "abhay" }
                },
                "required": ["name", "email", "password", "confirmPassword"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" },
                    "data": { "$ref": "#/components/schemas/UserData" }
                  },
                  "example": {
                    "message": "User created successfully",
                    "data": {
                      "_id": "681497b8457baeb8bb86eef3",
                      "name": "groot",
                      "email": "admin@gmail.com",
                      "role": "USER",
                      "pdf": ["https://cloudinary.com/xyz.pdf"],
                      "createdAt": "2025-05-02T10:00:24.141Z",
                      "updatedAt": "2025-05-02T10:00:24.141Z",
                      "__v": 0
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "User already exists or invalid data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": { "type": "string" }
                  },
                  "example": {
                    "message": "User already exists"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "UserData": {
        "type": "object",
        "properties": {
          "_id": { "type": "string", "example": "681497b8457baeb8bb86eef3" },
          "name": { "type": "string", "example": "groot" },
          "email": { "type": "string", "example": "admin@gmail.com" },
          "role": { "type": "string", "example": "USER" },
          "pdf": {
            "type": "array",
            "items": { "type": "string" },
            "example": ["https://cloudinary.com/xyz.pdf"]
          },
          "createdAt": { "type": "string", "format": "date-time", "example": "2025-05-02T10:00:24.141Z" },
          "updatedAt": { "type": "string", "format": "date-time", "example": "2025-05-02T10:00:24.141Z" },
          "__v": { "type": "integer", "example": 0 }
        }
      },
...
    }
  }
}